+++ doc/Changelog.md
@@ -1,3 +1,7 @@
+* mark media as selected when coming from single media
+* mark subalbum as selected when coming from it
+* make page scroll to subalbum when going up in album tree
+
 ### version v4.9.46 (Apr 23, 2021)
 
 * loading of albums with subalbums has been optimized